Add personal access token option in CLI
This option is interesting when no client_id and client_secret are generated from GitHub.
These usually are used for an application. In our case, we don't have an application, we
can generate a personal access token which uses HTTP Basic Auth. When used, the parameters
do not need to contain client id and params query information.